modified:   controllers/__pycache__/controller.cpython-313.pyc 	new file:   controllers/__pycache__/controller_match.cpython-313.pyc 	new file:   controllers/__pycache__/controller_round.cpython-313.pyc 	new file:   controllers/__pycache__/controller_tournament.cpython-313.pyc 	modified:   controllers/controller.py 	new file:   controllers/controller_match.py 	new file:   controllers/controller_round.py 	renamed:    tournaments/controller_tournament.py -> controllers/controller_tournament.py 	modified:   data_players.json 	deleted:    matchs/__pycache__/controller_match.cpython-313.pyc 	deleted:    matchs/controller_match.py 	modified:   models/__pycache__/model.cpython-313.pyc 	new file:   models/__pycache__/model_match.cpython-313.pyc 	new file:   models/__pycache__/model_round.cpython-313.pyc 	new file:   models/__pycache__/model_tournament.cpython-313.pyc 	modified:   models/model.py 	renamed:    matchs/model_match.py -> models/model_match.py 	new file:   models/model_round.py 	new file:   models/model_tournament.py 	deleted:    tournaments/__pycache__/controller_tournament.cpython-313.pyc 	modified:   tournaments/__pycache__/model_tournament.cpython-313.pyc 	deleted:    tournaments/__pycache__/view_tournament.cpython-313.pyc 	deleted:    tours/__pycache__/controller_round.cpython-313.pyc 	deleted:    tours/__pycache__/model_round.cpython-313.pyc 	deleted:    tours/__pycache__/view_round.cpython-313.pyc 	deleted:    tours/controller_round.py 	deleted:    tours/model_round.py 	modified:   views/__pycache__/view.cpython-313.pyc 	renamed:    matchs/__pycache__/view_match.cpython-313.pyc -> views/__pycache__/view_match.cpython-313.pyc 	new file:   views/__pycache__/view_round.cpython-313.pyc 	new file:   views/__pycache__/view_tournament.cpython-313.pyc 	renamed:    matchs/view_match.py -> views/view_match.py 	renamed:    tours/view_round.py -> views/view_round.py 	renamed:    tournaments/view_tournament.py -> views/view_tournament.py

+++ controllers/controller.py
@@ -2,11 +2,15 @@ import random
 import json
 from views.view import display_name, display_menu, display_lastname, display_date_of_birth, display_created_player, ask_player_creation
 from models.model import Player
-from tournaments.controller_tournament import create_tournament
-from matchs.controller_match import manage_winner_match, create_match
+from .controller_tournament import create_tournament
+from models.model_tournament import Tournament
+from .controller_match import manage_winner_match, create_match
+from models.model import Player
+from models.model_round import Round
 from pathlib import Path
 import os
-from tours.controller_round import start_round
+from .controller_round import start_round
+import datetime
 
 
 #CREATION DE JOUEUR
@@ -16,13 +20,13 @@ def create_player():
     last_name = display_lastname()
     date_of_birth = display_date_of_birth() 
     id = ""
-    player = Player(name=name,date_of_birth=date_of_birth,last_name=last_name,id=id)
+    player = Player(name=name,date_of_birth=date_of_birth,last_name=last_name,id=id, score=0)
     display_created_player(name, last_name)
     
     return player
 
 #SAUVEGARDE DU JOUEUR DANS UN FICHIER JSON
-def save_players(players):
+def save_players(players): # "players" contient la liste des joueurs qui vont être sauvegardés
     # sérialiser une **liste** complète (pas concaténer deux JSON)
     data = []
     for player in players: #Itération sur la liste contenant les infos des joueurs
@@ -31,18 +35,22 @@ def save_players(players):
         print("les données sauvegardées sont:", data_player)
         data.append(data_player) #Ajout des infos joueurs à la liste
     
-    with open('data_players.json', 'w', encoding='utf-8') as f: #Création d'un fichier JSON 
-        json.dump(data, f, ensure_ascii=False, indent=2) #Ecriture dans le fichier
+    
+    
+    if not os.path.exists("data_players.json"):
+        with open('data_players.json', 'w', encoding='utf-8') as f: #Création d'un fichier JSON 
+            json.dump(data, f, ensure_ascii=False, indent=2) #Ecriture dans le fichier
+    else: 
+        with open('data_players.json', 'a', encoding='utf-8') as f: #Création d'un fichier JSON 
+            json.dump(data, f, ensure_ascii=False, indent=2) #Ecriture dans le fichier
 
-def add_player_in_jsonfile():
-    pass
 
 
 #CONVERTISSEMENT D'UN OBJET PYTHON EN JSON
 def serializer(obj):
     """Convertit un objet Python en JSON"""
     if isinstance(obj, Player): #Vérification que l'objet de classe crée "obj" est bien du même type que Player
-        data_player = {"name": obj.name, "last_name":obj.last_name, "date_of_birth": obj.date_of_birth, "id":obj.id}
+        data_player = {"name": obj.name, "last_name":obj.last_name, "date_of_birth": obj.date_of_birth, "id":obj.id, "score":obj.score}
         print(f"Les données du joueur sont {data_player}")
         return data_player
     raise TypeError(f"Type non sérialisable: {type(obj)}")
@@ -74,10 +82,11 @@ def start_menu():
         IsCreation = True
         while IsCreation:
             player = create_player() #CREATION JOUEUR
+            print("Le type du joueur crée est:", type(player))
+            print()
             all_players.append(player)
-
-            if not os.path.exists("data_players.json"):
-                save_players(all_players) #CREATION D'UN FICHIER JSON
+            print("Le type du joueur présent dans la liste est:", type(all_players[0]))
+            save_players(all_players) #CREATION D'UN FICHIER JSON
             
             #Démarrage de la boucle pour création de joueur
             response = ask_player_creation()
@@ -96,28 +105,61 @@ def start_menu():
     
     #MATCH
     elif choice == 2:
-        
+        #play_match = Match()
         while len(all_players) <= 1:
             player = create_player()
             all_players.append(player)
         else:
-            create_match(all_players) #CREATION MATCH
-            manage_winner_match(all_players)
+            single_match = create_match(all_players) #CREATION MATCH
+            winner_match = manage_winner_match(single_match)
             
     #TOURNAMENT
     elif choice == 3:
-        create_tournament()
-        number_player_in_tournamment = input("Combien de joueurs participe au tournoi? ")
+        #CREATION DU TOURNOI
+        tournament = create_tournament() 
+        number_player_in_tournamment = int(input("Combien de joueurs participe au tournoi? ")) #Nombre de joueurs participant au tournoi
         
+        #CREATION DES JOUEURS DU TOURNOI 
         for player in range(number_player_in_tournamment):
             player = create_player()
-            all_players.append(player)
+            print()
+            all_players.append(player) #Enregistrement des joueurs dans la liste des joueurs sauvegardés
+        save_players(all_players) #Enregistrement des joueurs dans le fichier JSON
+        print()
+        
+        tournament.list_player_saved = all_players #Utilisation de la variable de l'objet Tournament
+        number_of_match = int(number_player_in_tournamment / 2) #Détermination du nombre de match
+        
+        round_match_list = [] #Création d'une liste contenant l'ensemble des matchs par round
+        winner_list = [] #Création d'une liste contenant l'ensemble des gagnants des matchs
+        short_lived_list = all_players.copy()
+        #Réalisation des matchs pour le round 1 
+        
+        for tour in range(tournament.number_of_round): #Execution des 4 rounds
+            Round.name_round = f"Round {tour+1}" #Nom du round
+            Round.date_and_hour_of_start = datetime.datetime.now()
+            tournament.actual_round = Round.name_round
+            print(f"Début du {Round.name_round} à {Round.date_and_hour_of_start}")
+            print()
+            short_lived_list = all_players.copy()
+            round_match_list = [] #Création d'une liste contenant l'ensemble des matchs par round
             
-            number_of_match = number_player_in_tournamment / 2
-            for match in number_of_match:
-                match = create_match()
+            for match in range(number_of_match):
+                match = create_match(short_lived_list) #Sélection aléatoire de 2 joueurs dans la liste "all_players " qui contient tous les joueurs
+                unique_match = ([match.players[0], match.score], [match.players[1], match.score])
+                round_match_list.append(unique_match) #Ajout du match qui vient d'être crée juste au-dessus dans la liste de tous les matchs du round
+                print()
+                print("Le match est:", unique_match)
+                winner = manage_winner_match(match.players) #Décide du vainqueur du match
+                print()
+                winner_list.append(winner) #Ajout du vainqueur dans une liste
+                print()
                 
-        start_round()
+            Round.date_and_hour_of_end = datetime.datetime.now()
+            print(f"Fin du {tournament.actual_round} à {Round.date_and_hour_of_end}")
+            #print("Voici les vainqueurs des différentes rencontres: ", winner_list)
+            
+
